Factor out common code and properly organize laws tests - Also enable Id Outcome laws
