+++ laws/shared/src/test/scala/cats/effect/laws/EitherTPureConcSpec.scala
@@ -15,37 +15,25 @@
  */
 
 package cats.effect
+package laws
 
-import cats.Show
 import cats.data.EitherT
-//import cats.laws.discipline.{AlignTests, ParallelTests}
+import cats.effect.testkit.{pure, PureConcGenerators, TimeT}, pure._, TimeT._
 import cats.laws.discipline.arbitrary._
-import cats.syntax.all._
-//import cats.effect.kernel.ParallelF
-import cats.effect.laws.GenTemporalTests
-import cats.effect.testkit._
-import cats.effect.testkit.TimeT._
-import cats.effect.testkit.{pure, PureConcGenerators}, pure._
 
-// import org.scalacheck.rng.Seed
-import org.scalacheck.util.Pretty
 import org.scalacheck.Prop
 
 import org.specs2.ScalaCheck
-// import org.specs2.scalacheck.Parameters
 import org.specs2.mutable._
 
 import scala.concurrent.duration._
 
 import org.typelevel.discipline.specs2.mutable.Discipline
 
-class EitherTPureConcSpec extends Specification with Discipline with ScalaCheck {
+class EitherTPureConcSpec extends Specification with Discipline with ScalaCheck with BaseSpec {
   import PureConcGenerators._
 
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit def execEitherT[E](sbool: EitherT[TimeT[PureConc[Int, *], *], E, Boolean]): Prop =
+  implicit def exec[E](sbool: EitherT[TimeT[PureConc[Int, *], *], E, Boolean]): Prop =
     Prop(
       pure
         .run(TimeT.run(sbool.value))
@@ -55,6 +43,5 @@ class EitherTPureConcSpec extends Specification with Discipline with ScalaCheck 
     "EitherT[TimeT[PureConc]]",
     GenTemporalTests[EitherT[TimeT[PureConc[Int, *], *], Int, *], Int]
       .temporal[Int, Int, Int](10.millis)
-    // ) (Parameters(seed = Some(Seed.fromBase64("IDF0zP9Be_vlUEA4wfnKjd8gE8RNQ6tj-BvSVAUp86J=").get)))
   )
 }
