+++ laws/shared/src/test/scala/cats/effect/laws/OptionTFreeSyncSpec.scala
@@ -17,17 +17,12 @@
 package cats.effect
 package laws
 
-import cats.{Eq, Eval, Show}
-import cats.free.FreeT
+import cats.{Eq, Eval}
 import cats.data.OptionT
+import cats.effect.testkit.{freeEval, FreeSyncEq, FreeSyncGenerators, SyncTypeGenerators}
+import cats.free.FreeT
 import cats.laws.discipline.arbitrary._
-import cats.effect.testkit.{freeEval, FreeSyncGenerators, SyncTypeGenerators}
-import cats.effect.testkit.FreeSyncEq
 import freeEval.{syncForFreeT, FreeEitherSync}
-import cats.syntax.all._
-
-import org.scalacheck.Prop
-import org.scalacheck.util.Pretty
 
 import org.specs2.ScalaCheck
 import org.specs2.mutable._
@@ -43,15 +38,6 @@ class OptionTFreeSyncSpec
 
   import FreeSyncGenerators._
   import SyncTypeGenerators._
-
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit val eqThrowable: Eq[Throwable] =
-    Eq.fromUniversalEquals
-
-  implicit def exec(sbool: FreeEitherSync[Boolean]): Prop =
-    run(sbool).fold(Prop.exception(_), b => if (b) Prop.proved else Prop.falsified)
 
   implicit val scala_2_12_is_buggy
       : Eq[FreeT[Eval, Either[Throwable, *], Either[Int, Either[Throwable, Int]]]] =
