+++ laws/shared/src/test/scala/cats/effect/laws/OptionTPureConcSpec.scala
@@ -15,50 +15,37 @@
  */
 
 package cats.effect
+package laws
 
-import cats.Show
-import cats.data.IorT
-//import cats.laws.discipline.{AlignTests, ParallelTests}
+import cats.data.OptionT
+import cats.effect.testkit.{pure, PureConcGenerators, TimeT}, pure._, TimeT._
 import cats.laws.discipline.arbitrary._
-import cats.syntax.all._
-//import cats.effect.kernel.ParallelF
-import cats.effect.laws.GenTemporalTests
-import cats.effect.testkit.{pure, PureConcGenerators}, pure._
-import cats.effect.testkit._
-import cats.effect.testkit.TimeT._
 
-// import org.scalacheck.rng.Seed
-import org.scalacheck.util.Pretty
 import org.scalacheck.Prop
 
 import org.specs2.ScalaCheck
-// import org.specs2.scalacheck.Parameters
 import org.specs2.mutable._
 
 import org.typelevel.discipline.specs2.mutable.Discipline
 
 import scala.concurrent.duration._
 
-class IorTPureConcSpec extends Specification with Discipline with ScalaCheck {
+class OptionTPureConcSpec extends Specification with Discipline with ScalaCheck with BaseSpec {
   import PureConcGenerators._
 
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit def execIorT[L](sbool: IorT[TimeT[PureConc[Int, *], *], L, Boolean]): Prop =
+  implicit def exec(sbool: OptionT[TimeT[PureConc[Int, *], *], Boolean]): Prop =
     Prop(
       pure
         .run(TimeT.run(sbool.value))
         .fold(
           false,
           _ => false,
-          iO => iO.fold(false)(i => i.fold(_ => false, _ => true, (_, _) => false)))
-    )
+          bO => bO.flatten.fold(false)(_ => true)
+        ))
 
   checkAll(
-    "IorT[PureConc]",
-    GenTemporalTests[IorT[TimeT[PureConc[Int, *], *], Int, *], Int]
+    "OptionT[TimeT[PureConc]]",
+    GenTemporalTests[OptionT[TimeT[PureConc[Int, *], *], *], Int]
       .temporal[Int, Int, Int](10.millis)
-    // ) (Parameters(seed = Some(Seed.fromBase64("IDF0zP9Be_vlUEA4wfnKjd8gE8RNQ6tj-BvSVAUp86J=").get)))
   )
 }
