+++ laws/shared/src/test/scala/cats/effect/laws/PureConcSpec.scala
@@ -15,35 +15,22 @@
  */
 
 package cats.effect
+package laws
 
-import cats.Show
-//import cats.laws.discipline.{AlignTests, ParallelTests}
+import cats.effect.testkit.{pure, PureConcGenerators, TimeT}, pure._, TimeT._
 import cats.laws.discipline.arbitrary._
-import cats.syntax.all._
-//import cats.effect.kernel.ParallelF
-import cats.effect.laws.GenTemporalTests
-import cats.effect.testkit._
-import cats.effect.testkit.TimeT._
-import cats.effect.testkit.{pure, PureConcGenerators}, pure._
 
-// import org.scalacheck.rng.Seed
-import org.scalacheck.util.Pretty
 import org.scalacheck.Prop
 
 import org.specs2.ScalaCheck
-// import org.specs2.scalacheck.Parameters
 import org.specs2.mutable._
 
 import scala.concurrent.duration._
 
 import org.typelevel.discipline.specs2.mutable.Discipline
 
-class PureConcSpec extends Specification with Discipline with ScalaCheck {
-//  import ParallelFGenerators._
+class PureConcSpec extends Specification with Discipline with ScalaCheck with BaseSpec {
   import PureConcGenerators._
-
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
 
   implicit def exec(fb: TimeT[PureConc[Int, *], Boolean]): Prop =
     Prop(pure.run(TimeT.run(fb)).fold(false, _ => false, _.getOrElse(false)))
@@ -51,11 +38,5 @@ class PureConcSpec extends Specification with Discipline with ScalaCheck {
   checkAll(
     "TimeT[PureConc]",
     GenTemporalTests[TimeT[PureConc[Int, *], *], Int].temporal[Int, Int, Int](10.millis)
-  ) /*(Parameters(seed = Some(Seed.fromBase64("OjD4TDlPxwCr-K-gZb-xyBOGeWMKx210V24VVhsJBLI=").get)))*/
-
-//  checkAll("PureConc", ParallelTests[PureConc[Int, *]].parallel[Int, Int])
-
-//  checkAll(
-//    "ParallelF[PureConc]",
-//    AlignTests[ParallelF[PureConc[Int, *], *]].align[Int, Int, Int, Int])
+  )
 }
