+++ laws/shared/src/test/scala/cats/effect/laws/ReaderWriterStateTFreeSyncSpec.scala
@@ -17,23 +17,19 @@
 package cats.effect
 package laws
 
-import cats.{Eq, Eval, Show}
-import cats.free.FreeT
-import cats.data.WriterT
-import cats.laws.discipline.arbitrary._
+import cats.{Eq, Eval}
+import cats.data.ReaderWriterStateT
 import cats.effect.testkit.{freeEval, FreeSyncGenerators, SyncTypeGenerators}
+import cats.free.FreeT
+import cats.laws.discipline.{arbitrary, MiniInt}, arbitrary._
 import freeEval.{syncForFreeT, FreeEitherSync}
-import cats.syntax.all._
-
-import org.scalacheck.Prop
-import org.scalacheck.util.Pretty
 
 import org.specs2.ScalaCheck
 import org.specs2.mutable._
 
 import org.typelevel.discipline.specs2.mutable.Discipline
 
-class WriterTFreeSyncSpec
+class ReaderWriterStateTFreeSyncSpec
     extends Specification
     with Discipline
     with ScalaCheck
@@ -42,20 +38,11 @@ class WriterTFreeSyncSpec
   import FreeSyncGenerators._
   import SyncTypeGenerators._
 
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit val eqThrowable: Eq[Throwable] =
-    Eq.fromUniversalEquals
-
-  implicit def exec(sbool: FreeEitherSync[Boolean]): Prop =
-    run(sbool).fold(Prop.exception(_), b => if (b) Prop.proved else Prop.falsified)
-
   implicit val scala_2_12_is_buggy
       : Eq[FreeT[Eval, Either[Throwable, *], Either[Int, Either[Throwable, Int]]]] =
     eqFreeSync[Either[Throwable, *], Either[Int, Either[Throwable, Int]]]
 
   checkAll(
-    "WriterT[FreeEitherSync]",
-    SyncTests[WriterT[FreeEitherSync, Int, *]].sync[Int, Int, Int])
+    "ReaderWriterStateT[FreeEitherSync]",
+    SyncTests[ReaderWriterStateT[FreeEitherSync, MiniInt, Int, MiniInt, *]].sync[Int, Int, Int])
 }
