+++ laws/shared/src/test/scala/cats/effect/laws/StateTFreeSyncSpec.scala
@@ -17,24 +17,19 @@
 package cats.effect
 package laws
 
-import cats.{Eq, Eval, Show}
-import cats.free.FreeT
-import cats.data.ReaderWriterStateT
-import cats.laws.discipline.arbitrary._
-import cats.laws.discipline.MiniInt
+import cats.{Eq, Eval}
+import cats.data.StateT
 import cats.effect.testkit.{freeEval, FreeSyncGenerators, SyncTypeGenerators}
+import cats.free.FreeT
+import cats.laws.discipline.{arbitrary, MiniInt}, arbitrary._
 import freeEval.{syncForFreeT, FreeEitherSync}
-import cats.syntax.all._
-
-import org.scalacheck.Prop
-import org.scalacheck.util.Pretty
 
 import org.specs2.ScalaCheck
 import org.specs2.mutable._
 
 import org.typelevel.discipline.specs2.mutable.Discipline
 
-class ReaderWriterStateTFreeSyncSpec
+class StateTFreeSyncSpec
     extends Specification
     with Discipline
     with ScalaCheck
@@ -43,20 +38,11 @@ class ReaderWriterStateTFreeSyncSpec
   import FreeSyncGenerators._
   import SyncTypeGenerators._
 
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit val eqThrowable: Eq[Throwable] =
-    Eq.fromUniversalEquals
-
-  implicit def exec(sbool: FreeEitherSync[Boolean]): Prop =
-    run(sbool).fold(Prop.exception(_), b => if (b) Prop.proved else Prop.falsified)
-
   implicit val scala_2_12_is_buggy
       : Eq[FreeT[Eval, Either[Throwable, *], Either[Int, Either[Throwable, Int]]]] =
     eqFreeSync[Either[Throwable, *], Either[Int, Either[Throwable, Int]]]
 
   checkAll(
-    "ReaderWriterStateT[FreeEitherSync]",
-    SyncTests[ReaderWriterStateT[FreeEitherSync, MiniInt, Int, MiniInt, *]].sync[Int, Int, Int])
+    "StateT[FreeEitherSync]",
+    SyncTests[StateT[FreeEitherSync, MiniInt, *]].sync[Int, Int, Int])
 }
