+++ laws/shared/src/test/scala/cats/effect/laws/WriterTPureConcSpec.scala
@@ -15,50 +15,40 @@
  */
 
 package cats.effect
+package laws
 
-import cats.Show
-import cats.data.OptionT
-//import cats.laws.discipline.{AlignTests, ParallelTests}
+import cats.data.WriterT
 import cats.laws.discipline.arbitrary._
-import cats.syntax.all._
-//import cats.effect.kernel.ParallelF
-import cats.effect.laws.GenTemporalTests
 import cats.effect.testkit._
 import cats.effect.testkit.TimeT._
 import cats.effect.testkit.{pure, PureConcGenerators}, pure._
 
-// import org.scalacheck.rng.Seed
-import org.scalacheck.util.Pretty
 import org.scalacheck.Prop
 
 import org.specs2.ScalaCheck
-// import org.specs2.scalacheck.Parameters
 import org.specs2.mutable._
-
-import org.typelevel.discipline.specs2.mutable.Discipline
 
 import scala.concurrent.duration._
 
-class OptionTPureConcSpec extends Specification with Discipline with ScalaCheck {
+import org.typelevel.discipline.specs2.mutable.Discipline
+
+class WriterTPureConcSpec extends Specification with Discipline with ScalaCheck with BaseSpec {
   import PureConcGenerators._
 
-  implicit def prettyFromShow[A: Show](a: A): Pretty =
-    Pretty.prettyString(a.show)
-
-  implicit def execOptionT(sbool: OptionT[TimeT[PureConc[Int, *], *], Boolean]): Prop =
+  implicit def exec[S](sbool: WriterT[TimeT[PureConc[Int, *], *], S, Boolean]): Prop =
     Prop(
       pure
-        .run(TimeT.run(sbool.value))
+        .run(TimeT.run(sbool.run))
         .fold(
           false,
           _ => false,
-          bO => bO.flatten.fold(false)(_ => true)
-        ))
+          pO => pO.fold(false)(p => p._2)
+        )
+    )
 
   checkAll(
-    "OptionT[TimeT[PureConc]]",
-    GenTemporalTests[OptionT[TimeT[PureConc[Int, *], *], *], Int]
+    "WriterT[PureConc]",
+    GenTemporalTests[WriterT[TimeT[PureConc[Int, *], *], Int, *], Int]
       .temporal[Int, Int, Int](10.millis)
-    // ) (Parameters(seed = Some(Seed.fromBase64("IDF0zP9Be_vlUEA4wfnKjd8gE8RNQ6tj-BvSVAUp86J=").get)))
   )
 }
